aiopsl drivers(DPNI, DPBP, DPCI, DPRC): port to MC 10.0.0
Updated AIOPSL drivers for MC objects(DPNI, DPBP, DPCI, DPRC) to be compatible with MC 10.0.0 ABI changes

+++ src/kernel/include/fsl_gen.h
@@ -47,7 +47,7 @@
 
 #define MAKE_UMASK32(_width)	(uint32_t)(((uint64_t)1 << (_width)) - 1)
 #define MAKE_UMASK64(_width)	((uint64_t)((_width) < 64 ? ((uint64_t)1 << (_width)) - 1 \
-                                 : -1))
+                                 : (uint64_t)-1))
 
 #define OS_MEM_RESERVED(start_addr,next_addr) volatile uint8_t reserved##start_addr [next_addr - start_addr] 
 /**< Reserved uint8_t space generation; Useful for memory maps */
